+++ api/src/index.ts
@@ -1,14 +1,16 @@
 import type express from 'express';
-import * as Sentry from "@sentry/node";
+import * as Sentry from '@sentry/node';
+import { type Server, createServer as createHttpServer } from 'http';
 import { createApp } from './app';
 import { addGraphqlRoutes } from './graphql/server';
 import { getChildLogger } from './logger';
-import { closeRedis } from './redis';
-import { type Server, createServer as createHttpServer } from 'http';
+import { closeRedis, initRedis } from './redis';
 import { PORT } from './env';
 import { errorHandler } from './middleware/errors';
-import { initRedis } from './redis';
-import "express-async-errors";
+import 'express-async-errors';
+import prisma from './database';
+import register from './metrics';
+import promBundle from 'express-prom-bundle';
 
 const logger = getChildLogger({ msgPrefix: 'SERVER' });
 
@@ -34,13 +36,17 @@ async function startServer(
   }
 
   app.use(errorHandler);
+  router.get('/metrics', async (req: express.Request, res: express.Response) => {
+    const prismaMetrics = await prisma.$metrics.prometheus();
+    const appMetrics = await register.metrics();
+    return res.send(prismaMetrics + appMetrics);
+  });
 
   const httpServer = createHttpServer(app);
   return httpServer.listen({ port: PORT }, () => {
     logger.info(`Running on ${PORT}`);
   });
 }
-
 
 /* istanbul ignore next @preserve */
 startServer(app, router)